LOGIN AND REGISTER DONE

+++ app/src/main/java/ni/edu/uam/finduam/model/UsuarioIdRequest.kt
@@ -1,5 +1,5 @@
 package ni.edu.uam.finduam.model
 
 data class UsuarioIdRequest(
-    val idUsuario: Int
+    val idUsuario: Long // Cambiado a Long para consistencia con la API
 )
